+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changes
 
 * [#550](https://github.com/clojure-emacs/cider-nrepl/pull/550): Always return test documentation messages as strings.
+* [#563](https://github.com/clojure-emacs/cider-nrepl/pull/563): Add :root-ex key to error summary that contains the classname of the root cause.
 
 ## 0.18.0 (2018-08-06)